Modificando variables
cambiamos variables enteras por tipo double

+++ Tecnicatura3Java/Lección09/Calculadora_UTN/src/CalculadoraUTN.java
@@ -38,10 +38,10 @@ public class CalculadoraUTN {
     private static void ejecutarOperacion(int operacion, Scanner entrada){
         // 9.3 Ponemos la entrada de información
         System.out.print("Ingrese el valor para el operando 1: ");
-        var operando1 = Integer.parseInt(entrada.nextLine());
+        var operando1 = Double.parseDouble(entrada.nextLine());
         System.out.print("Ingrese el valor para el operando 2: ");
-        var operando2 = Integer.parseInt(entrada.nextLine());
-        int resultado;
+        var operando2 = Double.parseDouble(entrada.nextLine());
+        double resultado;
     switch(operacion){
         case 1 -> { //Suma
             resultado = operando1 + operando2;
